use i1 = fp >> 32
On 64-bit platforms, this produces code with slightly better ILP:

	mov      e_i1,e_i2
	shr      $32,q_i1

The shift now does not depend on the move.  (The difference may be very
small on modern CPUs because register renaming is performed in the front-end,
that is, early enough.  The difference is measurable on Sandy Bridge.)

On 32-bit platforms, this means that i1 and i2 are now swapped.